# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Game Service including Player, ClawMachine, GachaMachine, and WhackAMole endpoints.",
        "title": "Game Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/clawMachine/addTouchedItemRecord": {
            "post": {
                "description": "Record an item that was touched during a claw game",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Add touched item record",
                "parameters": [
                    {
                        "description": "Add touched item record request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddTouchedItemRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Touched item record added successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/adjustPlayerCoin": {
            "post": {
                "description": "Add or subtract coins from a player's balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Adjust player coin balance",
                "parameters": [
                    {
                        "description": "Adjust coin request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdjustPlayerCoinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Player coin adjusted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/adjustPlayerDiamond": {
            "post": {
                "description": "Add or subtract diamonds from a player's balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Adjust player diamond balance",
                "parameters": [
                    {
                        "description": "Adjust diamond request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdjustPlayerDiamondRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Player diamond adjusted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/createClawItems": {
            "post": {
                "description": "Create multiple claw items with their properties",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Create claw items",
                "parameters": [
                    {
                        "description": "Claw items creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateClawItemsRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Claw items created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/createClawMachine": {
            "post": {
                "description": "Create a new claw machine with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Create a new claw machine",
                "parameters": [
                    {
                        "description": "Claw machine creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateClawMachineRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Claw machine created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/createClawPlayer": {
            "post": {
                "description": "Create a new claw player with initial coin and diamond balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Create a new claw player",
                "parameters": [
                    {
                        "description": "Claw player creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateClawPlayerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Claw player created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddTouchedItemRecordRequest": {
            "type": "object",
            "required": [
                "catched",
                "gameID",
                "itemID"
            ],
            "properties": {
                "catched": {
                    "type": "boolean"
                },
                "gameID": {
                    "type": "integer"
                },
                "itemID": {
                    "type": "integer"
                }
            }
        },
...
        "dto.AdjustPlayerCoinRequest": {
            "type": "object",
            "required": [
                "amount",
                "playerID",
                "type"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "playerID": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "plus",
                        "minus"
                    ]
                }
            }
        },
        "dto.AdjustPlayerDiamondRequest": {
            "type": "object",
            "required": [
                "amount",
                "playerID",
                "type"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "playerID": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "plus",
                        "minus"
                    ]
                }
            }
        },
...
        "dto.CreateClawItemsRequest": {
            "type": "object",
            "required": [
                "clawItems"
            ],
            "properties": {
                "clawItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreateClawItemRequest"
                    }
                }
            }
        },
...
    }
}